% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tableaus.R
\name{LP.Tableaus}
\alias{LP.Tableaus}
\title{LP.Tableaus}
\usage{
LP.Tableaus(dataframe1, varbase)
}
\arguments{
\item{varbase}{a vector}
}
\value{
a dataframe
}
\description{
If you know basic variables,you can get optimal tableu.
}
