% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LPSolver.R
\name{LPSolver}
\alias{LPSolver}
\title{LPSolver}
\usage{
LPSolver(LP, maxiter = 20, min = F, Bland = F)
}
\arguments{
\item{LP}{a list or a dataframe}
}
\value{
a list containing much information
}
\description{
Solve LP and show simplex tableaus!
}
\examples{
LP=list("max",c(4,2,3,4,2),c(1,2,1,0,0,1,2,2,0,0,0,0,0,1,1),c(8,8,3),"<,<,<")
r=LPSolver(LP)
}
