% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoptimal.R
\name{Simplex.isoptimal}
\alias{Simplex.isoptimal}
\title{LPtableaus}
\usage{
Simplex.isoptimal(LPtable, type = "max")
}
\arguments{
\item{LPtable}{dataframe}
}
\value{
True/False
}
\description{
main functions:LPSolver,LP.Tableaus,Simplex.dual,table2tex.Have fun!
}
