% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard.R
\name{Simplex.standard}
\alias{Simplex.standard}
\title{Simplex.standard <- function(orgin,M=NULL)}
\usage{
Simplex.standard(orgin, M = NULL)
}
\value{
list
}
\description{
find the standard form of a LP
}
\examples{
LP=list("max",c(4,2,3,4,2),c(1,2,1,0,0,1,2,2,0,0,0,0,0,1,1),c(8,8,3),"<,<,<")
r=Simplex.standard(LP)
}
